全ての主キーのデータ型をLong型に変更

+++ src/main/java/com/example/demo/repository/HistoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.History;
 
-public interface HistoryRepository extends JpaRepository<History, Integer> {
+public interface HistoryRepository extends JpaRepository<History, Long> {
 
 	List<History> findByAccountId(Long userId);
